+++ src/index.ts
@@ -1,2 +1,2 @@
-export { Money } from './money';
+export { Money, Money as default } from './money';
 export { IncompatibleCurrencyError, UnsafeIntegerError } from './errors';